Refactoring: moved ContentToolbar and all of its extensions and its inheritance to a own source file
