+++ taz.neo/ViewController/ContentVC.swift
@@ -49,103 +49,6 @@ public class ContentUrl: WebViewUrl, DoesLog {
   
 } // ContentUrl
 
-/// The ContentToolBar consists of a ToolBar and an encompassing view to position
-/// the toolbar with enough distance to the bottom safe area
-open class ContentToolbar: UIView {
-  
-  static let ToolbarHeight: CGFloat = 52
-  private var toolbar = Toolbar()
-  private(set) var heightConstraint: NSLayoutConstraint?
-
-  public var totalHeight: CGFloat {
-    return ContentToolbar.ToolbarHeight + UIWindow.bottomInset
-  }
-  
-  public override var backgroundColor: UIColor? {
-    didSet { toolbar.backgroundColor = self.backgroundColor }
-  }
-  
-  /// alpha for transluent color of the whole Toolbar Area
-  /// needed to ensure Home Indicator has same Background like Toolbar above
-  public var translucentAlpha: CGFloat {
-    get { return toolbar.translucentAlpha }
-    set { toolbar.translucentAlpha = newValue }
-  }
-  
-  public var translucentColor: UIColor {
-    get { return toolbar.translucentColor }
-    set { toolbar.translucentColor = newValue }
-  }
-  
-  public override init(frame: CGRect) {
-    super.init(frame: frame)
-    addSubview(toolbar)
-    pin(toolbar.top, to: self.top)
-    pin(toolbar.left, to: self.left)
-    pin(toolbar.right, to: self.right)
-    toolbar.pinHeight(ContentToolbar.ToolbarHeight)
-    toolbar.createBars(2)
-    self.clipsToBounds = true
-  }
-  
-  required public init?(coder: NSCoder) {
-    super.init(coder: coder)
-  }
-  
-  public func pinTo(_ view: UIView) {
-    view.addSubview(self)
-    pin(self.left, to: view.left)
-    pin(self.right, to: view.right)
-    pin(self.bottom, to: view.bottom)
-    heightConstraint = self.pinHeight(totalHeight)
-  }
-  
-  public func hide(_ isHide: Bool = true) {
-    if isHide {
-      UIView.animate(withDuration: 0.5) { [weak self] in
-        self?.heightConstraint?.isActive = false
-        self?.heightConstraint = self?.pinHeight(0)
-        self?.layoutIfNeeded()
-      }
-    }
-    else {
-      UIView.animate(withDuration: 0.5) { [weak self] in
-        self?.heightConstraint?.isActive = false
-        self?.heightConstraint = self?.pinHeight(self!.totalHeight)
-        self?.layoutIfNeeded()
-      }   
-    }
-  }
-  
-  public func addButton(_ button: ButtonControl, direction: Toolbar.Direction) {
-    toolbar.addButton(button, direction: direction)
-  }
-  
-  public func addArticleButton(_ button: ButtonControl, direction: Toolbar.Direction) {
-    toolbar.addButton(button, direction: direction, at: 1)
-  }
-  
-  public func addSectionButton(_ button: ButtonControl, direction: Toolbar.Direction) {
-    toolbar.addButton(button, direction: direction, at: 0)
-  }
-  
-  func setArticleBar() { toolbar.bar = 1 }
-  func setSectionBar() { toolbar.bar = 0 }
-  
-  public func setButtonColor(_ color: UIColor) { toolbar.setButtonColor(color) }
-  
-  public func setActiveButtonColor(_ color: UIColor) {
-    toolbar.setActiveButtonColor(color)
-  }
-  
-  public func applyDefaultTazSyle() {
-    self.setButtonColor(Const.Colors.iOSDark.secondaryLabel)
-    self.setActiveButtonColor(Const.Colors.ciColor)
-    self.backgroundColor = Const.Colors.iOSDark.secondarySystemBackground
-    self.translucentAlpha = 0.0
-  }
-}
-
 // MARK: - ContentVC
 /**
  A ContentVC is a view controller that displays an array of Articles or Sections
